Function Reference

LandUse.CParam — Type

A Country-wide Parameter struct

LandUse.Param — Type

Parameter Struct

This struct defines the parameter type for the model. It reads default values from the file src/params.json. Those values can be overridden by supplying a Dict to the constructor with key => value pairs:

p = LandUse.Param(par = Dict(:η => 0.1))

creates a Param instance with agglomeration forces set to 0.1

LandUse.Region — Type

Region type

Urban expansion model with flexible commuting cost and differential supply elasticity. This represents a single region. A region contains one potential urban area, and a rural area. A Region can be part of a wider Country.

LandUse.Urban — Type

Urban Model

This model has a fixed price of land ρr in the rural sector. Differently to our baseline model, where ρr is a GE object.

Base.convert — Method

convert a country param to one param for each region

LandUse.D — Method

Population Density at location $l$

LandUse.D2 — Method

Population Density at location $l$. second version, independent of Lu

LandUse.EqSys! — Method

Computes the entries of the residual vector $u$

LandUse.Eqsys! — Method

Eqsys!(F::Vector{Float64},m::Region,p::Param)

compute system of equations for the general (with flexible ϵ).

LandUse.Eqsys! — Method

Eqsys!(F::Vector{Float64},m::Urban,p::Param)

compute system of equations for the general (with flexible ϵ).

LandUse.GDP — Method

GDP

LandUse.H — Method

housing supply at location $l$

LandUse.Rmk — Method

Country Rural Market Clearing

LandUse.Rmk — Method

Rural Market Clearing

LandUse.Srh — Method

Residential Land in Rural sector. equation (20)

LandUse.Yr — Method

Production of Rural Good

LandUse.Yu — Method

Production of Urban Good

LandUse.cfun — Method

cost of construction at location $l$

LandUse.compute_ξw — Method

compute_ξw()

Computes input parameter ξw directly from data. Notice that $1 - \xi_w$ is the elasticity of speed to wage income. We measure from individual commuting data in ENL (Enquete National de Logement) an increase in commuting speed of about 11% from 1984 to 2013. This function returns

$\xi_w = 1 - \frac{11}{\%\Delta \theta_u}$

where $\%\Delta \theta_u$ stands for the percentage increase in urban wage from 1984 to 2013.

LandUse.copydb — Method

copy dropbox data to in/ folder

LandUse.cr — Method

optimal rural good consumption at location $l$. Equation (7) divided by p

LandUse.cs_plots — Method

single spatial cross sections region

LandUse.cu — Method

optimal urban good consumption at location $l$. Equation (8)

LandUse.cu_input — Method

Amount of Urban Good (numeraire) required to build housing at $l$

LandUse.cur — Method

urban good consumption in rural sector. Equation (8)

LandUse.d — Method

commuting distance

LandUse.dash — Method

dash(it;par = Dict())

Helper function for quick Region dashboard.

LandUse.dashboard — Method

dashboard(C::Vector{Country},it::Int)

Dashboard for a Country in period it

LandUse.dashboard — Method

dashboard(M::Vector{Region},p::Param,i::Int; objvalue = nothing)

Model dashboard for a single Region.

LandUse.dashk20 — Method

dashk20(;save = false)

Produce and optionally save model output for 20 city version.

LandUse.dataframe — Method

Obtain a Time Series from an array of Country as a DataFrame

LandUse.dataframe — Method

Obtain a Time Series for a single region as a DataFrame

LandUse.expmodel — Method

Estimate exponential model on spatial density.

LandUse.expmodel — Method

Compute exponential decay model of density for each region in a country

LandUse.firstnorm — Method

normalize a vector wrt it's first element

LandUse.foc_Lr — Method

FOC of rural firm wrt labor Lr

LandUse.foc_Sr — Method

FOC of rural firm wrt land Sr

LandUse.getfringe — Method

Get Fringe from indifference condition

At the fringe $\phi$ we have the condition

\[w(0) - \tau(\phi) = w_r\]

which can be rearranged to obtain a map from $w(0) - w_r = \tau(\phi)$.

The function takes $w(0) - w_r$ as argument x. then we give $\tau(\phi)$ to its inverse function to get back $\phi$

LandUse.i0 — Method

i0()

interact with the baseline single city version of the model. This basically calls dashboard(M::Vector{Region},p::Param,i::Int) after creating p from a set of user inputs.

LandUse.i0k — Method

i0k()

interact with the 2 city model.

LandUse.integrate! — Method

integrate!(m::Region,p::Param)

perform numerical integration using gauss-laguerre.

LandUse.integrate! — Method

integrate!(m::Urban,p::Param)

could be made much faster by filling a matrix col-wise with integration nodes and doing a matmul on it? have to allocate memory though.

LandUse.invτ — Method

inverse commuting cost. cost x → location. Notice we don't consider that cost is zero beyond ϕ: we want to find ϕ here to start with.

LandUse.iobj — Method

iobj()

Similar to i0 but prints the value of the objective function into one of the panels.

LandUse.isfeasible_grid_row — Method

checks whether a certain row of the grid together with a starting value x0 produces a valid solution in all periods.

LandUse.issue10 — Method

plot $\phi_k$ vs $L_{u,k}$ for all regions $k$ and different eps slopes. relates to https://github.com/floswald/LandUse.jl/issues/10

LandUse.issue11 — Method

measure increase in land value at fringe. by default starts in 1960 and measures increase up to 2020.

https://github.com/floswald/LandUse.jl/issues/11

LandUse.issue15 — Method

https://github.com/floswald/LandUse.jl/issues/15

LandUse.issue36 — Method

https://github.com/floswald/LandUse.jl/issues/36

same growth in sectors i. high cbar vs low sbar: show implied city density time series to see that only that config works ii. show falling housing spending share as well

implications of growth in either sector only

identify commuting cost params by matching time series data

LandUse.issue59 — Method

run model with flat epsilon

https://github.com/floswald/LandUse.jl/issues/59

LandUse.issue60 — Method

run model with agglomeration forces

https://github.com/floswald/LandUse.jl/issues/60

LandUse.issue64 — Method

run model with congestion forces

https://github.com/floswald/LandUse.jl/issues/64

LandUse.issue66 — Method

run multi city with congestion

https://github.com/floswald/LandUse.jl/issues/66

LandUse.issue67 — Method

five city cross section

https://github.com/floswald/LandUse.jl/issues/67

LandUse.issue9 — Method

plot $\phi_k$ vs $\L{u,k}$ for all regions $k$. relates to https://github.com/floswald/LandUse.jl/issues/9

LandUse.issue_21 — Function

https://github.com/floswald/LandUse.jl/issues/21

LandUse.issue_22 — Method

https://github.com/floswald/LandUse.jl/issues/22

LandUse.jc — Method

solve a Country at point x0

keywords:

estimateθ: whether the sequence of $\theta_u$ should be a choice variable

solve : solve the model (true) or return the model object for debugging?

fit_allyears : try to match the population distribution in all available years or just a single year.

LandUse.jm — Method

solve model at current Param p and starting at point x0

Keyword:

estimateθ: whether the sequence of $\theta_u$ should be a choice variable or not. default false.

LandUse.k20output — Method

Produces output from 20-city model. In particular comparison with single city case, and extension with d1, d2.

LandUse.k5 — Method

5 country case

LandUse.latex_param — Method

print default param to latex table

LandUse.maxnorm — Method

normalize a vector wrt it's maximal element

LandUse.multik — Method

Runs K regions as a country and produces 2 plots

LandUse.nearstart — Method

you just had a failed start at p

LandUse.objective — Method

objective(x; moments = false, plot = false, save = false, fname = "moments")

moment objective function for an optimizer

LandUse.objective1 — Method

objective1(;save = false)

run objective function at default parameter

LandUse.output_paper — Method

produces all output from the model needed to compile the paper.

This uses current baseline parameters defined in params.json

LandUse.p2Flux — Method

p2Flux(p::Param)

map param to Flux Chain to get starting value

LandUse.p2x — Method

p2x(p::Param)

map param to x for objective function quick eval this is the inverse of x2dict.

LandUse.pcy — Method

aggregate per capita income

LandUse.plot_dens_years — Method

plot densities for all years

LandUse.plotk20 — Method

plotk20(;save = false)

Produce model output for 20 city version.

LandUse.poparea_data — Method

load population and area data

LandUse.popdata_mapyears — Method

find closest year in population data to model years

LandUse.prepare_data — Method

prepare_data(p::Param; digits = 9)

Takes raw csv data and prepares to be used in model. writes to csv input files.

Smoothing of Productivity Series

We take the following steps to obtain a smoothed series for $\theta_r$ and $\theta_u$:

We obtain the estimated series at annual frequency.

We subset both series to start in 1840 and end in 2015 (rural productivity ends in that year)

We linearly interpolate the missing interwar years.

Smoothing is done via the [smooth] function from the QuantEcon package: we use the default Hann window and a 15-year window size. We experimented with the window size until high-frequency oscillations disappear.

Our rural productivity series gets very volatile from 2000 onwards, and in fact one would find a decreasing rural productivity series if we applied our smoother to post 2000 data. Therefore from the year 2000 onwards, we grow the smoothed series forward with 1% annual growth. Given that 2000 is very close to the final new steady state of the model, the actual choice of growth rate has only a small impact on our results.

LandUse.q — Method

house price function at $l$. equation (12)

LandUse.qr — Method

rural house price from land price

LandUse.rel_cityarea — Method

rel_cityarea(m::Model)

Computes city area relative to total rural land use. This is the model moment related to the data moment artificialized land relative to agricultural land in 2015. We measure this to be 18.3% in the data.

LandUse.relpop — Method

return model relative population in each year to largest city largest city is city 1

LandUse.run — Method

run Single region model for all time periods

LandUse.run — Method

run Single region model for all time periods

LandUse.runestim — Method

runestim(;steps = 1000,fname = "moments")

Run the default differential evolution optimizer from BlackBoxOptim.jl

LandUse.runk — Method

run Multi-region model for all time periods starting from the single city starting value.

LandUse.startvals_impl — Method

collect valid starting values for the k country case by reusing the first period solution of each preceding step in θu along the sequence

LandUse.startvals_k — Method

startvals_k()

Find feasible starting values for multi city case (2 to 5 cities) and write to disk. By default return the saved dict with start param and initial guess vector x0.

LandUse.startvals_par — Method

helper function to prepare country param

LandUse.targets — Method

returns a dict of empirical targets for the model

LandUse.ts_plots — Method

ts_plots(M,p::Param;fixy = false)

Time series plots for a single Region M.

LandUse.update! — Method

update!(c::Country,p::Vector{Param},x::Vector{Float64})

Update a Country object with a current vector of choice variables x supplied by the solver. The ordering in x is:

b: ratio of labor to land in region 1

r: land rent

pr: relative price rural good

4 : (K+3), Sr: amount of land use in rural production (complement of Srh)

(K+4) : (2K+3), Lu: urban pop in each k

(2K+4) : end, θu: urban prod in each k

LandUse.update! — Method

update!(m::Region,p::Param,x::Vector{Float64})

update a single Region with a Param at choices x.

LandUse.update! — Method

update!(m::Urban,p::Param,x::Vector{Float64})

update a single region a parameter vector at choices x.

LandUse.utility — Method

Utility function equation (5)

LandUse.w — Method

wage at location $l$

LandUse.wr — Method

rural wage from indifference condition at ϕ. Eq (11)

LandUse.wu — Method

urban wage at location $l$

LandUse.wu0 — Method

urban wage at center

LandUse.x2dict — Method

x2dict(x)

map vector x to dict for Param

LandUse.xsr — Method

excess subsistence rural worker

LandUse.xsu — Method

excess subsistence urban worker

LandUse.y — Method

disposable income: GDP net of commuting costs

LandUse.ρ — Method

land price function at $l$. equation (15)

LandUse.τ — Method

commuting cost: location x → cost

LandUse.χ — Method

housing supply shifter at $l$

LandUse.ϵ — Method

housing supply elasticity at $l$